Refactored folders, types and functions into more functional units as discussed in code review.
