--- package/src/values/animation/create/index.ts
@@ -1,2 +1,0 @@
-export * from "./createSpring";
-export * from "./createTiming";
