+++ package/src/values/animation/factory/index.ts
@@ -1,0 +1,2 @@
+export { createAnimationFactory } from "./createAnimationFactory";
+export * from "./types";
