+++ package/src/values/animation/factory/types.ts
@@ -1,0 +1,4 @@
+import type { AnimationParams } from "../types";
+
+export type RequiredAnimationParams = Required<Omit<AnimationParams, "from">> &
+  Pick<AnimationParams, "from">;
