+++ package/src/values/animation/functions/index.ts
@@ -1,5 +1,2 @@
-export * from "./bezier";
-export * from "./interpolate";
-export * from "./interpolateColors";
-export * from "./spring";
-export * from "./types";
+export { interpolate } from "./interpolate";
+export { interpolateColors, mixColors } from "./interpolateColors";
