--- package/src/values/animation/functions/types.ts
@@ -1,25 +1,0 @@
-export interface AnimationState {
-  duration: number;
-  startTime: number | null;
-  easing: EasingFunction;
-  done: boolean;
-  value: number;
-  from: number;
-  to: number;
-}
-
-export type EasingFunction = (value: number) => number;
-
-export interface TimingConfig {
-  from?: number;
-  to?: number;
-  duration?: number;
-  easing?: EasingFunction;
-}
-
-export interface SpringConfig {
-  mass?: number;
-  stiffness?: number;
-  velocity?: number;
-  damping?: number;
-}
