--- package/src/values/animation/hooks/index.ts
@@ -1,3 +1,0 @@
-export * from "./useTiming";
-export * from "./useLoop";
-export * from "./useSpring";
