+++ package/src/values/animation/index.ts
@@ -1,6 +1,3 @@
-export { runTiming, runSpring } from "./run";
-export { createTiming, createSpring } from "./create";
-export { useTiming, useSpring, useLoop } from "./hooks";
-export * from "./Easing";
-export * from "./Spring";
+export * from "./timing";
+export * from "./spring";
 export * from "./functions";
