--- package/src/values/animation/run/index.ts
@@ -1,2 +1,0 @@
-export * from "./runSpring";
-export * from "./runTiming";
