+++ package/src/values/animation/spring/functions/index.ts
@@ -1,0 +1,1 @@
+export { createSpringEasing } from "./spring";
