+++ package/src/values/animation/spring/functions/spring.ts
