+++ package/src/values/animation/spring/types.ts
@@ -1,0 +1,6 @@
+export interface SpringConfig {
+  mass?: number;
+  stiffness?: number;
+  velocity?: number;
+  damping?: number;
+}
