+++ package/src/values/animation/timing/Easing.ts
