+++ package/src/values/animation/timing/functions/__tests__/timing.spec.ts
