+++ package/src/values/animation/timing/functions/bezier.ts
