+++ package/src/values/animation/timing/functions/index.ts
@@ -1,0 +1,4 @@
+export { bezier } from "./bezier";
+export { timing } from "./timing";
+export * from "./types";
+export * from "./getResolvedParams";
