+++ package/src/values/animation/timing/functions/timing.ts
