+++ package/src/values/animation/timing/functions/types.ts
@@ -1,0 +1,8 @@
+export type EasingFunction = (value: number) => number;
+
+export interface TimingConfig {
+  from?: number;
+  to?: number;
+  duration?: number;
+  easing?: EasingFunction;
+}
